refactor: move the cache configuration and helpers into _preload.py
preload_and_cache.py becomes the compatibility re-export shim the plan calls
for (section 5, PR-21), and the machinery it held moves to the new private
module the mixin will join in the next commit: the CACHE[...] key-scheme
comment, the five lifetime/limit constants, and cache_lifetime_for_class,
is_preloading, pause_caching and resume_caching. Those 77 lines are lifted by
line span and are byte-identical.

All nine names preload_and_cache exports today still resolve from it, in the
PEP-484 redundant-alias form so they need neither a noqa nor a ratchet entry;
one line per name because ruff's isort runs with combine-as-imports off and
reports I001 for the parenthesized tuple. preload_and_cache.py has no per-file
ignores entry and does not gain one.

The keep-green edit this move requires: pdsfile.py imports the three names it
uses from ._preload, which is now canonical. pds3file and pds4file keep
importing cache_lifetime_for_class through the shim.
